==> 01_preparing_data.ipynb <==
Prepare data for generating synthetic population.

# Import libraries

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import json
import torch
import torch.nn as nn
from torch.nn import functional as F

# Global variables

MODEL_CONFIG = 'model/config.json'

DATA_PATH = 'data/raw_data.csv'

# ------------- Train, Test, Validation Set ----------- #
TEST_DF_PATH = 'data/test_df.csv'
TRAIN_DF_PATH = 'data/train_df.csv'
VALID_DF_PATH = 'data/valid_df.csv'

H_POP_PATH = 'data/h_population.csv'

# Util Functions

def plot_attribute_distribution(df, color=None):
    # check the distribution of each attribute
    # iterate over columns in dataframe
    for col in df.columns:
        # count the frequency of each attribute
        value_counts = df[col].value_counts()
        # create a bar chart of the frequency of each attribute
        plt.figure()
        if color is not None:
            plt.bar(value_counts.index, value_counts.values, color=color)
        else:
            plt.bar(value_counts.index, value_counts.values)
        plt.title(col)
        plt.show()

def print_uniq_val(df):
    # Create an empty dictionary to store unique values for each column
    unique_values_dict = {}

    # Iterate through each column of the dataframe
    for column_name in df.columns:
        # Get the unique values of the current column
        unique_values = df[column_name].unique().tolist()

        # Sort the unique values
        unique_values.sort()

        # Store the sorted unique values in the dictionary
        unique_values_dict[column_name] = unique_values

    # Print the unique values for each column
    for column_name, unique_values in unique_values_dict.items():
        print(f"Column name: {column_name}")
        print(f"Unique values: {unique_values}")
        print(f"Total unique values: {len(unique_values)}")
        print("\n")

def random_sample(df, n, random_state):
    """
    Returns a tuple of two dataframes:
    the first dataframe contains n randomly sampled rows from the input dataframe without replacement,
    the second dataframe contains the remaining rows that were not sampled.
    """
    sample_df = df.sample(n=n, replace=False, random_state=random_state)
    remaining_df = df.drop(sample_df.index)

    return sample_df, remaining_df

def count_unique_elements(my_list):
    # Create an empty list to store unique elements
    unique_list = []

    # Loop through each element in the list
    for element in my_list:
        # Check if the element is not already in the unique list
        if element not in unique_list:
            # If it's not, add it to the unique list
            unique_list.append(element)
            unique_list.sort()
            max_vocab = max(unique_list)

    # Return the length of the unique list and the unique list itself
    return len(unique_list), unique_list, max_vocab

# Import data

df = pd.read_csv(DATA_PATH)

df.info()

df.head()

ordered_col = ['sex', 'age', 'marst', 'classwk']
df = df[ordered_col]

df.head()

# Assigning new values to categorical variables



df['sex'] = df['sex'] + 200
df['marst'] = df['marst'] + 300
df['classwk'] = df['classwk'] + 400

df

# export the tokenized dataset
df.to_csv(H_POP_PATH, header=True, index=False)

# Split train, validation, test sets

len(df)

test_size = int(np.floor(0.1*len(df)))
remaining_size = len(df) - test_size
train_size = int(np.floor(0.8*remaining_size))
valid_size = remaining_size - train_size

print(f"test_size: {test_size}")
print(f"train_size: {train_size}")
print(f"valid_size: {valid_size}")

test_size + train_size + valid_size == len(df)

# randomly select "testing set" from the whole raw dataset
test_df, remaining_df = random_sample(df, n=test_size, random_state=2023)

test_df.info()

remaining_df.info()

# Split training and validation sets
train_df, valid_df = random_sample(remaining_df, n=train_size, random_state=2023)

train_df.info()

valid_df.info()

print(f'train size: {len(train_df)}')
print(f'validation size: {len(valid_df)}')
print(f'test size: {len(test_df)}')

# Save the testing, training, validation sets as CSV
test_df.to_csv(TEST_DF_PATH, header=True, index=False)

train_df.to_csv(TRAIN_DF_PATH, header=True, index=False)

valid_df.to_csv(VALID_DF_PATH, header=True, index=False)

# get the elements of the dataframe as a list
train_list = train_df.values.flatten().tolist()
valid_list = valid_df.values.flatten().tolist()

# print the resulting list
print("Example of the train list:")
print(train_list[:20])
print("#--------------------#")
print("Example of the valid list:")
print(valid_list[:20])
print("#--------------------#")


train_data = torch.tensor(train_list, dtype=torch.long)
val_data = torch.tensor(valid_list, dtype=torch.long)

print(train_data[-8:])
print(train_data[:8])
print(val_data[:8])
print(val_data[-8:])

# Vocab_size

# calculate the vocab_size for training model later on
vocab_size, unique_list, max_vocab = count_unique_elements(train_list)
print(f"vocab_size: {vocab_size}")
print(f"unique_list: {unique_list}")
print(f"maximum in vocab: {max_vocab}")

# Data to be written
dictionary = {
    "vocab_size": vocab_size,
    "max_vocab": max_vocab,
    "block_size": df.shape[1],
    "batch_size": 32,
	"eval_interval": 10,
	"learning_rate": 0.0001,
	"eval_iters": 10,
	"n_embd": 512,
	"n_head": 8,
	"n_layer": 6,
	"dropout": 0.1
}

# Serializing json
json_object = json.dumps(dictionary, indent=4)

# Writing to sample.json
with open(MODEL_CONFIG, "w") as outfile:
	outfile.write(json_object)
