==> 07_ctgan_training_generating.ipynb <==
# Install libraries

!pip install sdv

import sdv
print(sdv.__version__)

# !pip install sdmetrics

!pip install plotly>=5.10.0

!pip install -U kaleido

# Import libraries

from ctgan import CTGAN
import pandas as pd
from matplotlib import pyplot as plt
import numpy as np
# import sdv
# from sdv.tabular import CTGAN
# from sdv.evaluation import evaluate
from sdv.single_table import CTGANSynthesizer
from sdv.metadata import SingleTableMetadata
from sdmetrics.reports.single_table import QualityReport
import plotly
import json
import kaleido
import time

!python --version

pd.__version__

plotly.__version__

# Global variables

TRAIN_DF_PATH = 'data/train_df.csv'


# ------------- Model Training ------------- #
EPOCHS = 1000
BATCH_SIZE = 100

MODEL_PATH = 'model/ctgan.pkl'
SAVING_PATH = 'results/ctgan.csv'

# ------------- Generating Data ------------- #
GENERATING_SIZE = 7000

print(f'{MODEL_PATH = }')
print(f'{SAVING_PATH = }')

# Training phase

## Import Training Set

train_df = pd.read_csv(TRAIN_DF_PATH)

train_df.info()

train_df.head()

## Train model

# Names of the columns that are discrete
discrete_columns = ['sex', 'age2', 'edattain', 'classwk', 'marst', 'geo2_th2000']

metadata = SingleTableMetadata()
metadata.detect_from_dataframe(data=train_df)

synthesizer = CTGANSynthesizer(
    metadata=metadata, # required
    enforce_rounding=False,
    epochs=EPOCHS,
    verbose=True
)

# get the start time
st = time.time()

synthesizer.fit(train_df)

# get the end time
et = time.time()

# get the execution time
elapsed_time = et - st
print(f"Training time: {elapsed_time:.2f} seconds")

synthesizer.save(MODEL_PATH)

# Generating phase

# Load model
model = CTGANSynthesizer.load(MODEL_PATH)

# get the start time
st = time.time()

synthetic_data = model.sample(num_rows=GENERATING_SIZE)

# get the end time
et = time.time()

# get the execution time
elapsed_time = et - st
print(f"Generating time: {elapsed_time:.2f} seconds")

synthetic_data.head()

# Save the generated data
synthetic_data.to_csv(SAVING_PATH, index=False, header=True)
print(f'{SAVING_PATH = }')
